lambda/requirements.txt
attrs==21.2.0
aws-cdk-lib==2.0.0
aws-cdk.aws-lambda-python-alpha==2.0.0rc24
boto3==1.20.21
botocore==1.23.21
cattrs==1.8.0
constructs==10.0.10
jmespath==0.10.0
jsii==1.47.0
Pillow==8.4.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
